+++ mowl/base_models/__init__.py
@@ -1,4 +1,4 @@
-from .model import Model, EmbeddingModel
+from .model import Model
 from .elmodel import EmbeddingELModel
 from .graph_model import GraphModel, RandomWalkModel, KGEModel
 from .syntactic_model import SyntacticModel