+++ mowl/nn/el/elmodule.py
@@ -3,7 +3,9 @@ import torch.nn as nn
 
 
 class ELModule(nn.Module):
-    """Subclass of :class:`torch.nn.Module` for :math:`\mathcal{EL}` models. This class provides \
+    """Subclass of :class:`torch.nn.Module` for :math:`\mathcal{EL}` models. 
+
+    This class provides \
     an interface for loss functions of the 7 possible normal forms existing in the \
     :math:`\mathcal{EL}` language. In case a negative version of one of the loss function exist, \
     it must be placed inside the original loss function and be accesed through the ``neg`` \
@@ -37,7 +39,7 @@ class ELModule(nn.Module):
         """Loss function for GCI1: :math:`C_1 \sqcap C_2 \sqsubseteq D`.
 
         :param gci: Input tensor of shape \(\*,3\) where ``C1`` classes will be at \
-        ``gci[:,0]``, ``C2`` classes will be at gci[:,1] and ``D`` classes will be at \
+        ``gci[:,0]``, ``C2`` classes will be at ``gci[:,1]`` and ``D`` classes will be at \
         ``gci[:,2]``. It is recommended to use the :class:`ELDataset <mowl.datasets.el.ELDataset>`.
         :type gci: :class:`torch.Tensor`
         :param neg: Parameter indicating that the negative version of this loss function must be \
@@ -48,12 +50,11 @@ class ELModule(nn.Module):
         raise NotImplementedError()
 
     def gci2_loss(self, gci, neg=False):
-        """Loss function for GCI2: :math:`C \sqsubseteq R. D`.
-
+        """Loss function for GCI2: :math:`C \sqsubseteq \exists R.D`.
+        :math:`C \sqsubseteq \exists R. D`.
         :param gci: Input tensor of shape \(\*,3\) where ``C`` classes will be at \
         ``gci[:,0]``, ``R`` object properties will be at ``gci[:,1]`` and ``D`` classes will be \
-        at ``gci[:,2]``. It is recommended to use the \
-        :class:`ELDataset <mowl.datasets.el.ELDataset>`.
+        at ``gci[:,2]``. It is recommended to use the :class:`ELDataset <mowl.datasets.el.ELDataset>`.
         :type gci: :class:`torch.Tensor`
         :param neg: Parameter indicating that the negative version of this loss function must be \
         used. Defaults to ``False``.
@@ -63,7 +64,7 @@ class ELModule(nn.Module):
         raise NotImplementedError()
 
     def gci3_loss(self, gci, neg=False):
-        """Loss function for GCI3: :math:`R. C \sqsubseteq D`.
+        """Loss function for GCI3: :math:`\exists R.C \sqsubseteq D`.
 
         :param gci: Input tensor of shape \(\*,3\) where ``R`` object properties will be at \
         gci[:,0], ``C`` classes will be at ``gci[:,1]``  and ``D`` classes will be at \
@@ -94,7 +95,7 @@ class ELModule(nn.Module):
         """Loss function for GCI1 with bottom concept: :math:`C_1 \sqcap C_2 \sqsubseteq \perp`.
 
         :param gci: Input tensor of shape \(\*,3\) where ``C1`` classes will be at ``gci[:,0]``, \
-        ``C2`` classes will be at gci[:,1] and ``bottom`` classes will be at ``gci[:,2]``. It is \
+        ``C2`` classes will be at ``gci[:,1] and`` ``bottom`` classes will be at ``gci[:,2]``. It is \
         recommended to use the :class:`ELDataset <mowl.datasets.el.ELDataset>`.
         :type gci: :class:`torch.Tensor`
         :param neg: Parameter indicating that the negative version of this loss function must be \
@@ -105,7 +106,7 @@ class ELModule(nn.Module):
         raise NotImplementedError()
 
     def gci3_bot_loss(self, gci, neg=False):
-        """Loss function for GCI3 with bottom concept: :math:`R. C \sqsubseteq \perp`.
+        """Loss function for GCI3 with bottom concept: :math:`\exists R.C \sqsubseteq \perp`.
 
         :param gci: Input tensor of shape \(\*,3\) where ``R`` object properties will be at \
         gci[:,0], ``C`` classes will be at ``gci[:,1]``  and ``bottom`` classes will be at \
@@ -119,6 +120,13 @@ class ELModule(nn.Module):
         raise NotImplementedError()
 
     def get_loss_function(self, gci_name):
+        """
+        This chooses the corresponding loss fuction given the name of the GCI.
+
+        :param gci_name: Name of the GCI. Choices are ``gci0``, ``gci1``, ``gci2``, ``gci3``, \
+        ``gci0_bot``, ``gci1_bot`` and ``gci3_bot``.
+        :type gci_name: str
+        """
 
         if gci_name not in self.gci_names:
             raise ValueError(